refactor: remove `slot` module
